Include new version of data.json in source control
Please note that this is generated using a modified clear method which
deletes the nullified properties. This ensures that noise is filtered also
from the output besides the input is clean. There is an open PR for that
on the secret-handshake project.

This is not a source file as it is generated by other source files.
Inclusion of data.json in source control should be reconsidered.
An issue will be opened for this.